feat: reset project to clean start
Keep examples for ease of use for now
