+++ restaurantr/.gitignore
@@ -161,3 +161,8 @@ vendor/
 
 # CircleCI
 .circleci/generated_config.yml
+# @generated expo-cli sync-2b81b286409207a5da26e14c78851eb30d8ccbdb
+# The following patterns were generated by expo-cli
+
+expo-env.d.ts
+# @end expo-cli
