+++ restaurantr/app-example/(tabs)/explore.tsx
